For lists, use tab instead of space for indent.
This has no effect on rendered HTML.
However, it is a change that will allow for improved and correct
support of multiple nested lists with rich content inside.
In the end, this is a formatting change that "makes the most sense"
when considering all of Markdown features and how they interact.
It also more closely matches the original nested lists examples in
Markdown description (see
http://daringfireball.net/projects/markdown/syntax#list).