Remove Godeps file.
It wasn't being used, but rather included "just in case." It was out of
date. It's no longer needed as I trust latest versions of these
dependencies, and Travis should help catch problems if something breaks.
Conveniently, I need to push a commit to master to trigger initial
Travis build.