Use ANSI escape codes for bold text, when available.
The current implementation of terminalStringWidth is very hacky. I
could not come up with a better solution for now. I'll merge as is
because its behavior is helpful, and I don't expect too many other
styles to be added. Once a better upstream API is used, like
blackfriday v2, then perhaps this can be solved more elegantly. Until
then, this should suffice.

Resolves #25.