Refactor testcases
Refactor the test cases into input/output files to make it simpler to add
and modify them.

Test cases can now be defined by creating a file inside
`testdata/default` (for the default test suite) or by creating a new
directory in `testdata`. This new directory should contain a set of
input files (`*.in.md`) and matching output files (`*.out.md`) as well
as a single `options.json` that defines the set of options to invoke the
markdown formatter with.

Additionally adds new flag to the test cases called "-update_goldens"
that writes the matching `.out.md` file if the testcase fails.