bring back examples, simplify, port 2 more tests
This change brings back the two examples so they continue
to show up the package documentation.

It also removes a lot of functionality that is not used and not
justified at this time, given the basic testing requirements of
markdownfmt (i.e., no need to support different options).
It simplifies the code to a bare minimum (to make future
maintenance easier) and more consistent with my other code.

It ports two more tests, TestLineBreak and TestDoubleSpacedListEnd,
to use the new file-based test suite.