add minimal go.mod
The intention is to make it so that it's possible to import and use
this package in module mode (in addition to the well-supported
GOPATH mode), and have it build successfully. That doesn't happen
automatically because the latest version of the module that provides
the github.com/russross/blackfriday package (the v1 API) is an
v2.0.0+incompatible version that provides the v2 API, which is not
compatible.

Use the latest pseudo-version of blackfriday v1 API. Can't use the
latest tagged release version v1.5.2 because it doesn't include an
important bug fix for https://github.com/russross/blackfriday/issues/495.

Only specify the blackfriday version in go.mod; others are unspecified
to avoid the added burden of maintaining and updating them. The latest
version, whichever they are, are fully supported, just like in GOPATH
mode.

Fixes #47